Refactor round_float ray_submit check if restore folder exists
