+++ experiments/submissions.atari.yaml
@@ -1,0 +1,10 @@
+- Pong
+- Breakout
+- Space Invaders
+- Seaquest
+- Qbert
+# Other popular benchmark environments:
+- MsPacman
+- Enduro
+- Alien
+- BeamRider
