+++ experiments/submissions_pbt_grouped.yaml
@@ -401,19 +401,6 @@ pbt_num_envs_large:
         failures:
         - none
     (HumanoidStandup-v5):
-      tws47f25112513090df44:
-        status: FAILED
-        submission_id: pbt_num_envs_large_HumanoidStandup-v5_2541
-        failures:
-          ab237_00008:
-          - online history incomplete
-          ab237_00006:
-          - online history incomplete
-          ab237_00002:
-          - online history incomplete
-          ab237_00010:
-          - online history incomplete
-          multiple: online history incomplete
       tws47f251125175392044:
         status: SUCCEEDED
         submission_id: pbt_num_envs_large_HumanoidStandup-v5_9580
