+++ experiments/tune_parameters.json
@@ -17,11 +17,11 @@
       4.9e-07,
       1.55e-06,
       4.88e-06,
-      1.536e-05,
-      4.827e-05,
+      1.54e-05,
+      4.83e-05,
       0.0001,
-      0.00015177,
-      0.00047713,
+      0.000152,
+      0.000477,
       0.0015
     ]
   },
@@ -44,7 +44,7 @@
       1.0,
       10.0,
       40.0,
-      1000,
+      1000.0,
       null
     ]
   },
@@ -106,11 +106,11 @@
     "grid_search": [
       0.0,
       0.0001,
-      0.00031623,
+      0.000316,
       0.001,
-      0.00316228,
+      0.003162,
       0.01,
-      0.03162278,
+      0.031623,
       0.1,
       0.2,
       0.3162
@@ -121,26 +121,23 @@
       0.1,
       0.25,
       0.4,
-      0.5499999999999999,
+      0.55,
       0.7,
       0.85,
       1.0,
-      0.0,
       1.25,
       1.5
     ]
   },
   "vf_clip_param": {
     "grid_search": [
-      1,
-      5,
-      10,
-      25,
-      50,
-      75,
-      100,
-      500,
-      1000
+      0.05,
+      0.125,
+      0.33,
+      1.0,
+      2.0,
+      10.0,
+      100.0
     ]
   },
   "kl_coeff": {
